ramdisk: Reduce the cost of the interactive governor
Don't poll the cpu state as frquently.

Set a less power draining hispeed and touchboost freq
(msm8974 devices have found a moderate boost is all that
is needed to let interactive correctly handle these cases,
although the duration is still up for debate).

Do not use touchboost or gohispeed when in the powersave
profile.

Change-Id: Id7b1e5b156f0a31485bf1546c7139242e9b007c2