mofd-common: shims: Update SensorManager to work in cm-13.0
Shimming a constructor is very dangerous.  You don't know what
the actual object is!  In this case, our object got bigger
and therefore the previous attempt to shim it causes memory
corruption (see below).

Instead, provide a full implementation of this object.

This is based on the cm-12.1 source code modified to fix the
changes to other APIs for cm-13.0.

The memory corruption is caused by the fact that the legacy
blob knows the old size of the object and uses it to allocate
storage (stack or malloc) when creating objects.  When we shim
in the new object we allocate the old size but then call
the new constructor.

Compiling a simple test program in a cm-12.1 tree:

   SensorManager *s = new SensorManager();
   s->getDefaultSensor(0);

then running the test program with valgrind and the earlier shim
library (which mangles the symbols into the new object) reports:

==4374== Invalid write of size 4
==4374==    at 0x4CCCBFC: android::String16::String16(android::String16 const&) (in /system/lib/libutils.so)
==4374==    by 0x4A1E628: android::SensorManager::SensorManager(android::String16 const&) (SensorManager.cpp:92)
==4374==    by 0x4E4FBF6: android::SensorManager::SensorManager() (in /system/lib/libshim_camera_old.so)
==4374==    by 0x108506: ??? (in /system/bin/sensor-manager)
==4374==    by 0x1083FA: ??? (in /system/bin/sensor-manager)
==4374==    by 0x48A132D: __libc_init (in /system/lib/libc.so)
==4374==  Address 0x4e798d4 is 0 bytes after a block of size 36 alloc'd
==4374==    at 0x487DEA6: operator new(unsigned int) (vg_replace_malloc.c:327)
==4374==    by 0x1084FC: ??? (in /system/bin/sensor-manager)
==4374==    by 0x1083FA: ??? (in /system/bin/sensor-manager)
==4374==    by 0x48A132D: __libc_init (in /system/lib/libc.so)

Aka, the constructor is writing outside of the allocated size of
the object when it stores the (newly added) field for the
packageName which is what we'd expect.

Running the same test program with valgrind and this new
shim library reports no such problems.

Change-Id: Ic4b1e0ee923b576578b62447a225ab2e2a330f5a