mofd-common: init: Tweak telephony service startup
Drop the Asus hack of trying to restart the services with the
c_main service.  At least on M, that is firing off almost immediately
after the services are started anyway so it's really ugly.

As part of fixing that nasty hack, just start all the services.  We
are always DSDA anyway.

Let the nvmmanagers start as a core service and defer the other
services until late_start to ensure that there is plenty of
time for the nvmmanager to exist before they start.

Change-Id: Ie9645fed2b8dd7a143d829bd938c089f96c4a700