mofd-common: ril: responseDataCallListV6 status is a short
It appears that the hardware is actually sending back a short
that has been "copied" into the bytes of an int and not type
promoted to an int.

Specifically, I see "65535" as a status when the modem is not
properly attached to the network yet.

I think this is just a cosmetic change and doesn't affect any
behaviour, but it's nicer none the less.

Change-Id: I3083a5e2b63c2565fa50cd0b42aa5c9fda0a7329