mofd-common: ril: responseDataCallListV6 status is a short
It appears that the hardware is actually sending back a short
that has been "copied" into the bytes of an int and not type
promoted to an int.

Specifically, I see "65535" as a status when the modem is not
properly attached to the network yet.  Translating this into
-1 (PDP_FAIL_VOICE_REGISTRATION_FAIL) appears to fix a problem
where we end up "connected" but in a captive portal state.

We appear to end up in a captive portal state because f.o.telephony
is treating this as a successful connection but the radio did
not (could not) provide us with a DNS server which means we
are just plain hosed until you reset the radio (airplane) because
we think we are connected but we have no DNS servers configured.

Change-Id: I3083a5e2b63c2565fa50cd0b42aa5c9fda0a7329