mofd-common: Use ASUS_ZENFONE2_LP_BLOBS instead
The old name assumed that we were just making a few small changes
to handle gralloc incompatibility but it turns out that our
entire omx stack is painfully different from stock MM.

In at least one instance, Asus has added a field to a struct
that wasn't in 5.0 which makes this very specific to our device.

Use a global define instead of a board flag because the headers
used end up being included from several places and this ensures
that all headers include the correct definition, regardless of
where the #include is placed.

Change-Id: I013079d9a36de6d60afde87633dd64fb397effbc