mofd-common: ril: Rework power on hacks
The previous version never sent the response to the RADIO_POWER
command (oops, I meant to do a response at the end of the request
but screwed that up).

Fix that and rework the logic to set the automatic network selection
after the power on succeeds.  This gets rid of an incredibly ugly
"sleep(2)" put in there to ensure that the radio had time to
power itself back on successfully.

Change-Id: I85eb162d977091521e5a12c14c1b9542177eed54