mofd-common: init: Remove ondemand powersave bias
This was an attempt to work around the fact that ondemand was trying
too hard to avoid the lower frequencies.  This work around was only
partially effective and a better solution is to revert the revered
"remove the deadband effect" commit which really doesn't seem to make
a great deal of sense.

Change-Id: I057488c4939a53da6721363ef6d9ceded52829d7