mofd-common: overlay: Set our notification/battery led colours
Since we don't support blue, set all the values to be sure we
know what they are going to be.

The colours are unchanged from the current defaults except for
the notification which is changed from orange to green.

Allow users to disable the notification led and change its config.

Do not allow users to change the colours because we don't really
support much useful within the color palette.

Change-Id: I04fb891088d9edfb12c3bf6e75a98ca7595e3a51