mofd : redefine define audioserver
N brought in this change
https://github.com/CyanogenMod/android_frameworks_av/commit/dcb346b7dc5b88c3e85db8a70bbd6a2fee8192b9
this removes audioserver from mediaserver process, earlier in M we had redefined mediaserver because We need radio and/or system permissions to be able to talk to modem manager.
Without that communication, we don't know when there is a phone call
and therefore don't know when to connect the audio paths.

In N audioserver was moved out of mediaserver, so only redefining mediaserver doesn't solves our issue of talking to the modem, the audiosever does loads up but, after our modem has loaded this causes a failure in loading audiomanager for our modem(causes no audio in voice calls). Fix this by redefining audioserver and winning the race from modem.

Change-Id: Ifa0078b14568da779474e22c16b1bcb1324c85f5